% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heisertransformfunctions.R
\name{CreateVertices}
\alias{CreateVertices}
\title{CreateVertices Transforms categorical probabilities into vertices with x.y coordinates.}
\usage{
CreateVertices(df, var_left, var_top, var_right, verticeName = T)
}
\arguments{
\item{df}{The data frame where variables can be found.}

\item{var_left, var_top, var_right}{Columns in df that contain
probabilities per row, use "columnnames", names (left, top, right)
refer to place in final triangle.}

\item{verticeName}{Do you want to return the names of the columns in the
final data frame?.}
}
\description{
Select your data frame and the three variables with probabilities.
The returning dataframe will have the x and y coordinates of the 3 points
that represent the vertices.
Plot the vertices and use Prob2coord() to create the xy coordinates for all the rows.
}
\examples{
\dontrun {CreateVertices(testdata, "test1", "test2","test3",verticeName = F)}
}
\seealso{
Other Heiser.transform.functions: \code{\link{Prob2Coord}}
}

