% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heisertransformfunctions.R
\name{Prob2Coord}
\alias{Prob2Coord}
\title{Transforms categorical probabilities of every row into x.y coordinates.}
\usage{
Prob2Coord(df, var_left, var_top, var_right, append = FALSE)
}
\arguments{
\item{df}{The data frame where variables can be found.}

\item{var_left, var_top, var_right}{Columns in df that contain
probabilities per row, use "columnnames", names (left, top, right)
refer to place in final triangle.}

\item{append}{Return with or without original data frame?.}
}
\description{
Select your data frame and the three variables with probabilities.
The returning dataframe will have the x and y coordinates for every row.
When used in conjuction with CreateVertices() plot the vertices
and add the points.
}
\examples{
\dontrun {Prob2Coord(testdata, "test1", "test2","test3",append = TRUE)}
}
\seealso{
Other Heiser.transform.functions: \code{\link{CreateVertices}}
}

