% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heisertransform.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{testdata.}
\format{A data frame with 11 observations on the following 4 variables. \code{test1}, \code{test2},
  \code{test3}, \code{cat}}
\usage{
testdata
}
\description{
Full baby name data provided by the SSA. This includes all names with at
least 5 uses.
}
\keyword{datasets}

